backend/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==3.2.16
django-filter==23.4
django-templated-mail==1.1.1
djangorestframework==3.12.4
djoser==2.2.2
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
gunicorn==20.1.0
idna==3.6
isort==5.12.0
oauthlib==3.2.2
Pillow==9.3.0
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
